+++ test/Precept.Tests/TypeChecker/TypeCheckerFieldDefaultTests.cs
@@ -21,11 +21,60 @@ public class TypeCheckerFieldDefaultTests
     }
 
     [Fact]
+    public void FieldDefault_IntegerForQuantityField_EmitsTypeMismatch()
+    {
+        var precept = """
+            precept Widget
+            field q as quantity of 'mass' default 5
+            state Open initial
+            """;
+
+        TypeCheckerTestHelpers.CheckExpectingError(precept, DiagnosticCode.TypeMismatch);
+    }
+
+    [Fact]
+    public void FieldDefault_ComputedFieldRefCrossDimension_EmitsDimensionCategoryMismatch()
+    {
+        var precept = """
+            precept Widget
+            field source as quantity of 'length' default '5 m'
+            field target as quantity of 'mass' <- source
+            state Open initial
+            """;
+
+        TypeCheckerTestHelpers.CheckExpectingError(precept, DiagnosticCode.DimensionCategoryMismatch);
+    }
+
+    [Fact]
     public void FieldDefault_MatchingDimension_NoDiagnostic()
     {
         var precept = """
             precept Widget
             field q as quantity of 'mass' default '5 kg'
+            state Open initial
+            """;
+
+        TypeCheckerTestHelpers.CheckExpectingClean(precept);
+    }
+
+    [Fact]
+    public void FieldMin_WrongDimension_EmitsDiagnostic()
+    {
+        var precept = """
+            precept Widget
+            field q as quantity of 'mass' min '0 m'
+            state Open initial
+            """;
+
+        TypeCheckerTestHelpers.CheckExpectingError(precept, DiagnosticCode.InvalidTypedConstantContent);
+    }
+
+    [Fact]
+    public void FieldMax_MatchingDimension_NoDiagnostic()
+    {
+        var precept = """
+            precept Widget
+            field q as quantity of 'mass' max '100 kg'
             state Open initial
             """;